Extract out register and login into single auth form component.
