+++ client/src/pages/About_RHESSys.tsx
@@ -56,7 +56,10 @@ const SidePanel = () => {
   const navigate = useNavigate();
   return (
     <>
-      <BackButton onClick={() => navigate({ to: "/about" })} label="Back to about page" />
+      <BackButton
+        onClick={() => navigate({ to: "/about" })}
+        label="Back to about page"
+      />
       <Typography component="h2" variant="h2" fontWeight="bold" marginY={2}>
         About RHESSys
       </Typography>
