+++ client/src/pages/Scenarios.tsx
@@ -36,7 +36,10 @@ const SidePanel = () => {
   const navigate = useNavigate();
   return (
     <>
-      <BackButton onClick={() => navigate({ to: "/about" })} label="Back to about page" />
+      <BackButton
+        onClick={() => navigate({ to: "/about" })}
+        label="Back to about page"
+      />
       <Typography component="h2" variant="h2" fontWeight="bold" marginY={2}>
         Model Scenarios
       </Typography>
